About
Closed Updated 13 minutes ago

The Office of the National Coordinator for Health Information Technology (ONC), in collaboration with the Chamber of Digital Commerce announces a health standards and data innovation Code-a-Thon. Code-a-Thon contestants will use open source distributed ledger technology and health-specific standards to address one of the following health IT Tracks:

Identity Management and APIs: Demonstrate the use of blockchain to support the identity management requirements for APIs for both provider-to-provider and patient-mediated exchange.

Metadata Tagging and Policy Expression: Demonstrate the use of blockchain for security metadata and tagging to manage access, provide auditing and provenance information.

Data Aggregation and Linkage: Use of blockchain technology to securely aggregate, link and share patient data to the destination of their choice.

Concepts should be innovative. This should be a forward-thinking project that enhances interoperability and focuses on demonstrating the potential to seamlessly incorporate blockchain solutions into existing health IT Systems.

Contestants will access open data sources for test data and leverage openly available blockchain and health IT tools and resources.